init option instead
