+++ packages/paperclip-repl/package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "yarn build:definition && webpack",
+    "build:watch": "concurrently \"yarn build:definition\" \"webpack serve\"",
+    "build:definition": "paperclip build --only=d.ts"
   },
   "author": "",
   "license": "ISC",
@@ -15,12 +17,19 @@
     "paperclip-loader": "^18.2.14",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "redux": "^4.1.2",
+    "redux-saga": "^1.1.3",
     "tandem-designer": "^18.2.14"
   },
   "devDependencies": {
+    "concurrently": "^6.5.1",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.5.0",
     "mini-css-extract-plugin": "^2.4.5",
-    "ts-loader": "^9.2.6"
+    "ts-loader": "^9.2.6",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.7.1"
   }
 }
