+++ packages/paperclip-repl/paperclip.config.json
@@ -1,0 +1,27 @@
+{
+  "srcDir": "./src",
+  "lintOptions": {
+    "enforceVars": [
+      "font-family",
+      "color",
+      "background",
+      "background-color",
+      "margin",
+      "margin-left",
+      "margin-right",
+      "margin-top",
+      "margin-bottom",
+      "padding",
+      "padding-left",
+      "padding-right",
+      "padding-top",
+      "padding-bottom",
+      "border",
+      "border-left",
+      "border-right",
+      "border-top",
+      "border-bottom"
+    ]
+  },
+  "moduleDirs": ["./node_modules", "../../node_modules"]
+}
